+++ RELEASE_NOTES.md
@@ -22,6 +22,12 @@ This version also adds memory optimized emoji keyboard styles, which makes the e
 * `EmojiKeyboardStyle` has new memory `.optimized` styles, which are applied by default.
 * `Keyboard+StorageValue` is a new type that is used to persist codable types in storage.
 
+### 💡 Renamings
+
+* `SystemKeyboard` has been renamed to `KeyboardView`.
+* `SystemKeyboardBottomRow` has been renamed to `Keyboard.BottomRow`.
+* `SystemKeyboardButtonPreview` has been renamed to `Keyboard.ButtonPreview`.
+
 ### 💡 Adjustments
 
 * `EmojiKeyboard` renders better on iPad devices.